requirements.txt
flask>=2.0.0
boto3>=1.28.0
botocore>=1.31.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
requests>=2.31.0
uuid>=1.30
typing>=3.7.4
